perf(loops): cast flatfield-corrected fields to float32
Flatfield division (array / flatfield_dict[ch]) promotes uint16 to
float64, so the stitched canvas and every array derived from it were
float64 — double the necessary size. On long multi-channel timelapses
(e.g. 72 h, 4 channels) that canvas is the dominant host-RAM cost and was
triggering SLURM oom_kill during loading/stitching, before segmentation.

Cast the corrected field to float32 at the source. A 24-bit mantissa far
exceeds the precision of 16-bit camera data, and it halves the resident
canvas (~300-400 GB → ~150-200 GB on the 72 h case). Downstream
(scale_img, gamma, Cellpose, feature extraction) all handle float32.

test_loops.py: assert _load_well_fields returns float32 stacks.

+++ src/omero_screen/loops.py
@@ -514,7 +514,14 @@ def _load_well_fields(
                     f"Channel '{ch}' not found in flatfield correction masks. "
                     f"Available channels: {list(flatfield_dict.keys())}."
                 )
-            img = array[..., ch_idx] / flatfield_dict[ch]
+            # Flatfield division promotes uint16 → float64. Cast straight to
+            # float32: a 24-bit mantissa is far more precision than 16-bit
+            # camera data carries, and it halves the resident size of the
+            # stitched canvas and every array derived from it — the dominant
+            # host-RAM cost on long multi-channel timelapses.
+            img = (array[..., ch_idx] / flatfield_dict[ch]).astype(
+                np.float32, copy=False
+            )
             # Reduce (tzyx) → (tyx)
             img = np.squeeze(img, axis=1)
             per_channel[ch].append(img)
